Codegen: replaced C++20 std::format with my own Formatter class
It allowed to reduce compiler requirements from C++20 (and in particular, gcc 13+) to C++17, and potentially to any C++11 compiler (once we will replace std::string_view with easypb::string_view). It also reduced codegen.exe size by 40 KB and compile time from 4 to 3 sec.

The implemented Formatter class only replaces "{}" and "{\d}" with strings, but it's just 60 LOC. The formatting procedure divided into the two stages - first, the template string compiled to the vector of literal+argnum pairs (which is done only once for each template string), and then the compiled template applies to the actual parameters. This makes the formatting operation as fast as the bunch of strcat operations.